requirements.txt
appdirs==1.4.4
asgiref==3.5.0
atomicwrites==1.4.0
attrs==21.4.0
black==21.7b0
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==0.5.0
Django==3.0.3
django-cors-headers==3.7.0
django-filter==2.3.0
django-templated-mail==1.1.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
filelock==3.6.0
gunicorn==20.0.4
identify==2.4.11
idna==3.3
iniconfig==1.1.1
isort==5.5.2
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.1.0
more-itertools==8.12.0
mypy-extensions==0.4.3
nodeenv==1.6.0
oauthlib==3.2.0
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.1
pluggy==0.13.1
pre-commit==2.13.0
psycopg2==2.8.6
py==1.11.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
pytest==6.0.1
pytest-django==3.9.0
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
PyYAML==6.0
regex==2022.3.2
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.3
uritemplate==4.1.1
urllib3==1.26.8
virtualenv==20.13.3
whitenoise==5.2.0
